Squashed commit of the following:
commit 7bf05f502b958b60e3cfd4a0ce937fb19de57051
Merge: 1b24c68 8fbaa38
Author: Cathan Bertram <55544660+CathanBertram@users.noreply.github.com>
Date:   Wed Jan 26 11:17:01 2022 +0000

    Merge pull request #13 from lukewhitingdev/Cathan-Bertram

    Item System

commit 8fbaa380b55b67da43b73be86abc846be776c654
Merge: d5567e6 fcc5a45
Author: Cathan Bertram <55544660+CathanBertram@users.noreply.github.com>
Date:   Wed Jan 26 10:59:45 2022 +0000

    Merge branch 'Cathan-Bertram' of https://github.com/lukewhitingdev/Necrodoggiecon into Cathan-Bertram

commit d5567e6413c3aede8921fc2290b3a48f56595e9b
Author: Cathan Bertram <55544660+CathanBertram@users.noreply.github.com>
Date:   Wed Jan 26 10:59:31 2022 +0000

    better dynamic_cast cache

commit fcc5a4599589bf99fe1bcb4ee6a637c8041f9388
Merge: 13141b6 1b24c68
Author: Cathan Bertram <55544660+CathanBertram@users.noreply.github.com>
Date:   Wed Jan 26 10:47:56 2022 +0000

    Merge branch 'main' into Cathan-Bertram

commit 13141b6d504fa56e751015f9a0a1e8e0bdfda7a2
Author: Cathan Bertram <55544660+CathanBertram@users.noreply.github.com>
Date:   Wed Jan 26 10:44:32 2022 +0000

    cached dynamic_cast result

commit 1b24c68ea528dcd80aadf6cead07b2b39de9ca6d
Merge: 3d8d14d 2bca460
Author: Luke Whiting <43708922+lukewhitingdev@users.noreply.github.com>
Date:   Wed Jan 26 10:16:02 2022 +0000

    Merge pull request #10 from lukewhitingdev/Luke-Whiting

    Luke whiting - 3D Audio

commit 3d8d14d70ddf9a36479c8153296d321acc29f24c
Merge: 80c65cd 43f6c66
Author: Luke Whiting <43708922+lukewhitingdev@users.noreply.github.com>
Date:   Wed Jan 26 10:12:20 2022 +0000

    Merge pull request #12 from lukewhitingdev/bug-create-entity-inside-update-crash

    Bug-Fix for adding entities inside an entity's update loop crash

commit 6ac29f13fe6c4dd689cb321f6f65fc8792dd6a6b
Merge: aba9329 80c65cd
Author: Cathan Bertram <55544660+CathanBertram@users.noreply.github.com>
Date:   Wed Jan 26 09:40:59 2022 +0000

    Merge branch 'main' into Cathan-Bertram

commit 43f6c66292d21a103d043ed45fd8039c7aefd145
Author: Luke Whiting <43708922+lukewhitingdev@users.noreply.github.com>
Date:   Tue Jan 25 17:34:58 2022 +0000

    Fix for bug where creating a entity inside another entity would crash the program

commit 80c65cd014e93c4d7542cd000ff5e62da0fd6c03
Author: Luke Whiting <43708922+lukewhitingdev@users.noreply.github.com>
Date:   Tue Jan 25 17:24:06 2022 +0000

    Rollback to commit: https://github.com/lukewhitingdev/Necrodoggiecon/commit/6dce1a5e18f44ff22e5284cc42e105681d045b66

commit aba9329157ef2cd9c3137185b35831b81d80d9d5
Author: Cathan Bertram <55544660+CathanBertram@users.noreply.github.com>
Date:   Tue Jan 25 17:08:48 2022 +0000

    slight changes

commit e48e4992a58b2a63583fa77420121af2781ae0a3
Merge: 6dce1a5 e6eb2a7
Author: Ben Brown <49481217+Padingy@users.noreply.github.com>
Date:   Tue Jan 25 16:21:11 2022 +0000

    Merge pull request #11 from lukewhitingdev/Ben-Brown

    Collision Detection

commit e6eb2a77151bd0e47e5f24d4d8ed21661b3f584c
Merge: 999224d 6dce1a5
Author: Ben Brown <49481217+Padingy@users.noreply.github.com>
Date:   Tue Jan 25 16:20:44 2022 +0000

    Merge branch 'main' into Ben-Brown

commit 999224dd43fa070a7bd44fae447256c151b7b56f
Author: Ben Brown <49481217+Padingy@users.noreply.github.com>
Date:   Tue Jan 25 16:01:33 2022 +0000

    Finished Collision

    Only bit left is class specific HasCollided functions

commit 7b8d677e7610ae35f07ba5ffc188b74cca9f9e6c
Author: Cathan Bertram <55544660+CathanBertram@users.noreply.github.com>
Date:   Tue Jan 25 14:43:22 2022 +0000

    Item system seems to work

commit 1e825eaa2616891cc1d412a211b89c249fbedbf3
Merge: ab14d34 dd5a21e
Author: Ben Brown <49481217+Padingy@users.noreply.github.com>
Date:   Tue Jan 25 13:07:02 2022 +0000

    Merge branch 'main' into Ben-Brown

commit ab14d341f1959e623458d431af3654579ccb5cc0
Author: Ben Brown <49481217+Padingy@users.noreply.github.com>
Date:   Tue Jan 25 13:01:35 2022 +0000

    Save

commit 62d7ebc1567f24e3a10d8bcd12be04313e24f8b9
Merge: 4733f23 b0f530c
Author: Cathan Bertram <55544660+CathanBertram@users.noreply.github.com>
Date:   Tue Jan 25 08:30:07 2022 +0000

    Merge branch 'main' into Cathan-Bertram

commit 4733f237f73b1bff6e6e9f89a36182f4e86f2da8
Author: Cathan Bertram <55544660+CathanBertram@users.noreply.github.com>
Date:   Mon Jan 24 21:49:59 2022 +0000

    item stuff half ish done

commit 7d829d56fa5b28d4c0b55bda1326b60a15f19a32
Author: Ben Brown <49481217+Padingy@users.noreply.github.com>
Date:   Fri Jan 21 16:39:59 2022 +0000

    Basic Collision Stuff

commit ac200d6663841188a36c721b9a95c0a7dd764baa
Merge: 206908b 09311c4
Author: Ben Brown <49481217+Padingy@users.noreply.github.com>
Date:   Fri Jan 21 09:30:45 2022 +0000

    Merge branch 'main' into Ben-Brown

commit 206908b4fd5bb831f35638c2aa922ee053e5c125
Author: Ben Brown <49481217+Padingy@users.noreply.github.com>
Date:   Fri Jan 21 03:24:29 2022 +0000

    Changes to switch PCs

commit e860ce0a8ef75e6500dce305d865880c595752d7
Merge: c383202 7962352
Author: Ben Brown <49481217+Padingy@users.noreply.github.com>
Date:   Thu Jan 20 09:53:45 2022 +0000

    Merge branch 'main' into Ben-Brown

commit c3832021981430f7aedd8622e88f3bf907aacea0
Author: Ben Brown <49481217+Padingy@users.noreply.github.com>
Date:   Thu Jan 20 09:51:16 2022 +0000

    Commit to merge main into branch

commit 79b4f8c70c7e2dad25f1a5b2ac9e941b947c9fe8
Author: Ben Brown <49481217+Padingy@users.noreply.github.com>
Date:   Tue Jan 18 16:31:44 2022 +0000

    Some Maths and Development

commit 977ab9c7754f03d4ea259a5acee89047b2b0779d
Author: Ben Brown <49481217+Padingy@users.noreply.github.com>
Date:   Fri Jan 14 16:24:54 2022 +0000

    Class files made

+++ Cerberus/CCharacter.h
@@ -13,7 +13,7 @@ protected:
 	void AddVerticalMovement(int dir, float speed, float deltaTime);
 	void AddHorizontalMovement(int dir, float speed, float deltaTime);
 public:
-	void ApplyDamage(int damageAmount, CEntity* damageCauser) { OnTakeDamage(damageAmount, damageCauser); }
+	void ApplyDamage(float damageAmount, CEntity* damageCauser) { OnTakeDamage(damageAmount, damageCauser); }
 
 };
